deleted dead handlerinstructions
